requirements.txt
clipspy
streamlit